Bypass project XSD validation on Mono
Due to an apparent bug in Mono's XSD validation, types specified with xsi:type="Foo" are not being namespace-qualified, even though there is a default namespace specified in the XML root element. So instead of searching the XSD for the namespace-qualified type {http://www.sil.org/CogProject/1.0:Aline}, the Mono XSD validator is searching for the unqualified type {Aline}, and not finding it.

To bypass this bug, we just skip project-file validation entirely if we're running on the Mono platform.