New helper function: get new (empty) project
Since Cog configuration lives in project files, so that a project file is a
configuration file and vice-versa, we'll need to be able to create a blank
project file. This NewProject.cogx file is the same one that is used by
Cog.Application, but moved into the Cog.Application.CommandLine namespace
so that it can compile on Linux. (Cog.Application references GraphSharp,
which references some assemblies from WPF that aren't available in Mono).

In the future, we'll allow config files to be specified on the command line.