New command: syllabify
Syllabification takes as input either a list of words (one per line), or a
list of pre-syllabified words with periods as syllable separators and,
optionally, pipe characters as stem markers (before and after the stem).

The output of the syllabification command is suitable for feeding into the
make-pairs command, in preparation for the distance command.