First implementation of distance command
Output score is currently not very useful; what we probably want will be
a number between 0.0 and 1.0, where 1.0 means "very distant" and 0.0 means
identical. What we currently have is an integer where the higher the score,
the closer the words. But this is a good first step towards getting a proper
distance score later.