Bugfix for Alignment.ToString with no notes
If the notes parameter to Alignment.ToString(IEnumerable<string> notes) was
an empty collection, Cog would trow an exception when notesList was later
indexed. The code is supposed to add empty notes to the notesList until it
has as many entries as the alignment's ColumnCount, but that was incorrectly
skipped if the alignment had no notes at all. Fixed by simply removing the
if (hasNotes) check, so that notesList will always be populated.