Refactor common code into base class
The commands that need to create a new CogProject were all using the same
(duplicated) code to do so; move that into the base class. It looks like it
should go into the constructor, but in the future we'll need to parse the
command-line options before we can know what project to use. So we have to
use an explicit method call for setting up the project.