Bugfix: use only one alignment, no matter what
Words with two alignments of equal weight were printing out twice, once per
possible alignment. If the two alignments were of unequal weight, only the
best one was chosen, so the easiest way to see this bug was to pick two
words of a single character each. Bug is now fixed and a regression test
has been added to the unit test suite as well.