Refactor: Move common code into base class
All the FooVerb classes had the same structure in DoWork(): set up
an Errors object, populate it with error messages if any, then check it
at the end (and set retcode non-zero if there were any errors). When
all your derived classes have a common structure, that's a sign that
that code should be in the base class.