Bugfix: Don't throw exception if output pipe closed
On Mono (but not on Windows), trying to write to stdout when it's piped
into another process that has already exited would throw an exception:

System.IO.IOException: Write fault on path /home/username/src/cog/[Unknown]
  at System.IO.FIleStream.WriteInternal (...)
  at System.IO.FileStream.Write (...)
  at System.IO.StreamWriter.Flush (...)
  at System.IO.StreamWriter.Dispose (...)
  at System.IO.TextWriter.Dispose ()
  at SIL.Cog.Application.CommandLine.CommonOptions.RunAsPipe (...)

The best way I can find to handle this is to catch the IOException, check
if it's a "Write fault on path" exception, and ignore only that condition.
This might fail if the user's UI language is not English, but since
IOException objects don't carry a culture-invariant way of checking what
they are, we have to do it the not-quite-perfect way.