Use three digits for version numbers, not four
The Windows version installer uses names like Cog-1.1.0-Setup.exe with
just three version numbers, not four. Debian considers 1.1.0.0 to be
"after" 1.1.0, so if we later want to switch to four version numbers,
we'll be able to. But now is the time to switch to three numbers, before
we've shipped to any users.

Also changing the Debian part of the version number to -0alpha1 so that
once we release, we can bump it to -1 properly.