Build script changes: use MSBuild, not shell
Instead of a make-deb.sh script, we will now make the .deb file using
nothing but MSBUild task steps. A few shell commands are necessary, like
calling dpkg-parsechangelog to get the version number of the previous
build, but those are pretty self-contained.

In addition to replacing make-deb.sh, this commit also adds the Cog
config file to the files built into the Debian package. We will use it
later on, to allow the user to tweak parameters to some of Cog's
internal operations like word distance.