Add new command-line options for config files
New options:
 * --config-file (or -c for short) to use a file
 * --config-data (no shorthand) to use XML data passed directly as string

If neither of these are passed, the config will be searched for in the
standard Linux locations (as per the XDG spec): .config/, and then
/usr/local/share and /usr/share (NOT /etc, because our "config" file is
really a data file, and so it lives in /usr/share/cog-cmdline).

This commit also includes unit tests for both options (config file as file,
and as string passed on the command-line).