Fix XdgDirectoryTests on Linux
Some Linux distributions such as Linux Mint add paths to $XDG_DATA_DIRS
besides the defaults specified in the XDG spec. We shouldn't expect that
the value of XdgDirectories.DataDirs or ConfigDirs will be exactly what
the spec calls for, only that the appropriate data directories (/usr/share
and /usr/local/share) or config directories (/etc/xdg) will be included.