Fix build issues with .deb file
The .deb file was being built with a different version in the
debian/control and debian/changelog files; fix that so that the version
that the build process adds to the changelog file will be the same
version as the one that it adds to the control file.

Also update the timestamp in debian/changelog to match the actual time
of the initial release.