Rewrite "git describe" parsing script in Python
Since we want to use the same build process in both Linux and Windows
builds, our script to parse the output of "git describe" should not be a
shell script (nor should it be a Windows batch file). Since Python is a
standard part of our build process for other Windows builds, we can make
it a standard part of our build process here as well.