Fix race condition in event firing
If another thread deregistered the last event handler between the null check
and firing the event, you can end up trying to call a null handler. The
solution is to make a local copy of the event handler, and check the local
copy before calling it.

For more details, see:
http://blogs.msdn.com/b/ericlippert/archive/2009/04/29/events-and-races.aspx